Updated merging script

+++ CheckPointing/fastaStats_driver.sh
@@ -1,14 +1,11 @@
 #!/bin/bash
 #Script to perform merge mergedfasta files and retain only
 #the specified unique data (by sequence, ID, or both)
-#Usage: bash fastaStats_driver.sh sortedFolder genotypes
-#Usage Ex: bash fastaStats_driver.sh sortedCoordinate_samtoolsHisat2_run1 Y05 Y023_5 E05 R2 PA Sierra
-#Usage Ex: bash fastaStats_driver.sh sortedCoordinate_samtoolsTophat2_run1 Y05 Y023_5 E05 R2 PA Sierra
-#Usage Ex: bash fastaStats_driver.sh trimmed_run1 Y05 Y023_5 E05 R2 PA Sierra
-#Default usage Ex: bash fastaStats_driver.sh assemblyTrinity_all
-
-#Load module necessary for crc servers
-module load R
+#Usage: bash fastaMerge_driver.sh mergeBy sortedFolder genotypes
+#Usage Ex: bash fastaMerge_driver.sh sequence sortedCoordinate_samtoolsHisat2_run1 Y05 Y023_5 E05 R2 PA Sierra
+#Usage Ex: bash fastaMerge_driver.sh sequence sortedCoordinate_samtoolsTophat2_run1 Y05 Y023_5 E05 R2 PA Sierra
+#Usage Ex: bash fastaMerge_driver.sh sequence trimmed_run1 Y05 Y023_5 E05 R2 PA Sierra
+#Default usage Ex: bash fastaMerge_driver.sh sequence assemblyTrinity_all
 
 #Check for input arguments of folder names
 if [ $# -eq 0 ]; then
@@ -20,77 +17,91 @@ fi
 counter=1
 fastaList=""
 
+#Retrieve fasta output absolute path
+outputsPath=$(grep "mergedFASTA:" ../InputData/outputPaths.txt | tr -d " " | sed "s/mergedFASTA://g")
 #Retrieve fasta file path
 inputsPath=$(grep "assembling:" ../InputData/outputPaths.txt | tr -d " " | sed "s/assembling://g")
 
 #Determine assembly target
-if [[ "$1" == sorted* ]]; then
-	#Set output directory
-	outputFolder="$inputsPath/$1""_assemblyGenomeTrinity_mergedFasta"
+if [[ "$2" == sorted* ]]; then
+	#Create output directory
+	outputFolder="$outputsPath/$2""_assemblyGenomeTrinity_merged"$1"Fasta"
 	#Set output file names
 	mergedFastaFile="$outputFolder/Trinity.fasta"
-	summaryFile="$outputFolder/$1""_assemblyGenomeTrinity_mergedFasta_summary.txt"
+	summaryFile="$outputFolder/$2""_assemblyGenomeTrinity_merged"$1"Fasta_summary.txt"
 	#Retrieve selected fasta files
 	#Loop through all input genotypes and merge fasta files
 	for i in "$@"; do
-		#Skip first argument
+		#Skip first two arguments
 		if [[ $counter -eq $# ]]; then
 			#Add fasta file to list
-			fastaFile="$inputsPath/$1$i""_assemblyGenomeTrinity/Trinity.fasta"
+			fastaFile="$inputsPath/$2$i""_assemblyGenomeTrinity/Trinity.fasta"
 			fastaList="$fastaList$fastaFile"
-		elif [[ $counter -ge 2 ]]; then
+		elif [[ $counter -ge 3 ]]; then
 			#Add fasta file to list
-			fastaFile="$inputsPath/$1$i""_assemblyGenomeTrinity/Trinity.fasta "
+			fastaFile="$inputsPath/$2$i""_assemblyGenomeTrinity/Trinity.fasta "
 			fastaList="$fastaList$fastaFile "
 		fi
 		counter=$(($counter+1))
 	done
-elif [[ "$1" == trimmed* ]]; then
-	#Set output directory
-	outputFolder="$inputsPath/$1""_assemblyTrinity_mergedFasta"
+elif [[ "$2" == trimmed* ]]; then
+	#Create output directory
+	outputFolder="$outputsPath/$2""_assemblyTrinity_merged"$1"Fasta"
 	#Set output file names
 	mergedFastaFile="$outputFolder/Trinity.fasta"
-	summaryFile="$outputFolder/$1""_assemblyTrinity_mergedFasta_summary.txt"
+	summaryFile="$outputFolder/$2""_assemblyTrinity_merged"$1"Fasta_summary.txt"
 	#Retrieve selected fasta files
 	#Loop through all input genotypes and merge fasta files
 	for i in "$@"; do
-		#Skip first argument
+		#Skip first two arguments
 		if [[ $counter -eq $# ]]; then
 			#Add fasta file to list
-			fastaFile="$inputsPath/$1$i""_assemblyTrinity/Trinity.fasta"
+			fastaFile="$inputsPath/$2$i""_assemblyTrinity/Trinity.fasta"
 			fastaList="$fastaList$fastaFile"
-		elif [[ $counter -ge 2 ]]; then
+		elif [[ $counter -ge 3 ]]; then
 			#Add fasta file to list
-			fastaFile="$inputsPath/$1$i""_assemblyTrinity/Trinity.fasta "
+			fastaFile="$inputsPath/$2$i""_assemblyTrinity/Trinity.fasta "
 			fastaList="$fastaList$fastaFile "
 		fi
 		counter=$(($counter+1))
 	done
 else #Default accept a list of full file paths
-	#Set output directory
-	outputFolder="$inputsPath/$1_mergedFasta"
+	#Create output directory
+	outputFolder="$outputsPath/$2_merged"$1"Fasta"
 	#Set output file names
 	mergedFastaFile="$outputFolder/Trinity.fasta"
-	summaryFile="$outputFolder/$1""_mergedFasta_summary.txt"
+	summaryFile="$outputFolder/$2""_merged"$1"Fasta_summary.txt"
 	#Retrieve selected fasta files
 	#Loop through all input genotypes and merge fasta files
 	for i in "$@"; do
-		#Skip first argument
+		#Skip first two arguments
 		if [[ $counter -eq $# ]]; then
 			fastaList="$fastaList$i"
-		elif [[ $counter -ge 2 ]]; then
+		elif [[ $counter -ge 3 ]]; then
 			fastaList="$fastaList$i "
 		fi
 		counter=$(($counter+1))
 	done
 fi
+#Check if the folder already exists
+mkdir "$outputFolder"
+if [ $? -ne 0 ]; then
+	echo "The $outputFolder directory already exsists... please remove before proceeding."
+	exit 1
+fi
+
+#Merge the set of fasta files
+bash fastaMerge_driver.sh $1 $mergedFastaFile $fastaList
+
+#Write fasta stats to the summary file
+bash fastaStats.sh $mergedFastaFile $fastaList > $summaryFile
 
 #Write fasta stats to the csv formatted summary file
 summaryFileCSV=$(echo "$summaryFile" | sed 's/\.txt/\.csv/g')
-bash fastaStats_csvFormatted.sh $fastaList $mergedFastaFile > $summaryFileCSV
+bash fastaStats_csvFormatted.sh $mergedFastaFile $fastaList > $summaryFileCSV
 
 #Plot fasta stats from summary file
-Rscript fastaStats_barPlot.r $summaryFileCSV $1
+Rscript fastaStats_barPlot.r $1 $summaryFileCSV
 
 #Clean up
 rm Rplots.pdf
